temp windows fixes

+++ onlineupdate/source/libmar/src/mar_read.c
@@ -333,7 +333,7 @@ int get_mar_file_info_fp(FILE *fp,
     }
   }
 
-  if (ftell(fp) == offsetToContent) {
+  if (ftell(fp) == (long)offsetToContent) {
     *hasAdditionalBlocks = 0;
   } else {
     if (numAdditionalBlocks) {
